逆向工程/skills/reverse_indicator_id/requirements.txt
dashscope>=1.20.0
python-dotenv>=1.0.0
pyyaml>=6.0
